refactor(core): split "core" words further
Seriously, core is all you need to compile new words, but It's going to suck to program with.
